requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
anndata==0.7.6
antlr4-python3-runtime==4.8
anyio==3.3.1
argon2-cffi==20.1.0
astroid==2.7.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
black==21.9b0
bleach==3.3.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==2.0.5
click==8.0.1
configparser==5.0.2
cycler==0.10.0
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.2
docker-pycreds==0.4.0
einops==0.3.2
entrypoints==0.3
filelock==3.0.12
flake8==3.9.2
Flask==2.0.1
fsspec==2021.8.1
future==0.18.2
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.40.0
h5py==3.4.0
hydra-core==1.1.1
identify==2.2.15
idna==3.2
importlib-resources==5.2.2
iniconfig==1.1.1
ipykernel==6.0.2
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.9.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-server==1.11.0
jupyterlab==3.1.12
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.1
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
llvmlite==0.37.0
louvain==0.7.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
mpmath==1.2.1
multidict==5.1.0
mypy==0.910
mypy-extensions==0.4.3
natsort==7.1.1
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
nodeenv==1.6.0
nose==1.3.7
notebook==6.4.0
numba==0.54.0
numexpr==2.7.3
numpy==1.20.3
oauthlib==3.1.1
omegaconf==2.1.1
packaging==21.0
pandas==1.3.3
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.9.0
pathtools==0.1.2
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
platformdirs==2.3.0
pluggy==1.0.0
pre-commit==2.15.0
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.19
protobuf==3.18.0
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyDeprecate==0.3.1
pyflakes==2.3.1
Pygments==2.9.0
pylint==2.10.2
pynndescent==0.5.4
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
python-igraph==0.9.6
pytorch-lightning==1.4.7
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rpy2==3.4.5
rsa==4.7.2
scanpy==1.8.1
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.7.1
sentry-sdk==1.3.1
shortuuid==1.0.1
sinfo==0.3.4
six==1.16.0
smmap==4.0.0
sniffio==1.2.0
statsmodels==0.12.2
stdlib-list==0.8.0
subprocess32==3.5.4
sympy==1.8
tables==3.6.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
texttable==1.6.4
threadpoolctl==2.2.0
toml==0.10.2
tomli==1.2.1
torch==1.9.0
torchaudio==0.9.0
torchmetrics==0.5.1
torchtext==0.10.0
torchtyping==0.1.4
torchvision==0.10.0
tornado==6.1
tqdm==4.62.2
traitlets==5.0.5
typeguard==2.12.1
typing-extensions==3.10.0.2
tzlocal==3.0
umap-learn==0.5.1
urllib3==1.26.6
virtualenv==20.6.0
wandb==0.12.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==1.2.0
yapf==0.31.0
yarl==1.6.3
yaspin==2.1.0
zipp==3.5.0
